zng_time

Struct INSTANT_APP

pub struct INSTANT_APP;

App control of the INSTANT service in an app context.

Implementations

impl INSTANT_APP

pub fn set_mode(&self, mode: InstantMode)

Set how the app controls the time.

If mode is set to InstantMode::Now the custom now is unset.

pub fn set_now(&self, now: DInstant)

Set the INSTANT.now for the app threads.

Panics

Panics if the mode is InstantMode::Now.

pub fn advance_now(&self, advance: Duration)

Set the INSTANT.now for the app threads to the current time plus advance.

Panics

Panics if the mode is not InstantMode::Manual.

pub fn unset_now(&self)

Unset the custom now value.

pub fn custom_now(&self) -> Option<DInstant>

Gets the custom now value.

This value is returned by INSTANT.now if set.

pub fn pause_for_update(&self) -> Option<InstantUpdatePause>

If mode is InstantMode::UpdatePaused sets the app custom_now to the current time and returns an object that unsets the custom now on drop.